// onboarding/src/utils/validators/onboarding.schema.ts
export const deliveryPartnerPersonalSchema = z.object({
  have: z.string().min(1, { message: errorMessage('Yes or No') }),
  name: z.string().optional(),
  rate: z.string().optional(),
});
// ...
export const step6Schema = z.object({
  personal: deliveryPartnerPersonalSchema,
  partner: z.object({
    email: validateEmail,
    password: z.string().min(1, { message: errorMessage('Password') }),
    // warehouses: z.array(deliveryPartnerWarehouseSchema),
  }),
});

export type Step6Schema = z.infer<typeof step6Schema>;